// src/api/model/fixtureCard.ts
/**
 * Generated by orval v7.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FixtureGroupRef } from './fixtureGroupRef';
import type { UnitEnum } from './unitEnum';
import type { FixtureTypeEnum } from './fixtureTypeEnum';
import type { FixtureCardUnit } from './fixtureCardUnit';

export interface FixtureCard {
  /** @nullable */
  id?: number | null;
  fixtureCode: string;
  fixtureName: string;
  fixtureGroup: FixtureGroupRef;
  defaultUnit: UnitEnum;
  fixtureType: FixtureTypeEnum;
  /** @nullable */
  optimalLevel?: number | null;
  /** @nullable */
  minimumLevel?: number | null;
  /** @nullable */
  specialCode?: string | null;
  /** @nullable */
  shelfLocation?: string | null;
  /** @nullable */
  fixtureCardUnits?: FixtureCardUnit[] | null;
}